Updationg resolvers path folders
