+++ angular/wolox-books/src/app/app.module.ts
@@ -22,8 +22,8 @@ import { InterceptorService } from './interceptors/app.interceptor';
 import { AuthGuard } from './guards/auth.guard';
 import { UnauthGuard } from './guards/unauth.guard';
 
-import { GetBookListResolver } from './services/book/resolvers/get-book-list.resolver';
-import { GetBookDetailResolver } from './services/book/resolvers/get-book-detail.resolver';
+import { GetBookListResolver } from './screens/auth/screens/books/resolvers/get-book-list.resolver';
+import { GetBookDetailResolver } from './screens/auth/screens/book-detail/resolvers/get-book-detail.resolver';
 
 @NgModule({
   declarations: [
