+++ angular/wolox-books/src/app/screens/auth/screens/book-detail/book-detail.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 
-import { Book } from './../../../../models/book.model';
+import { Book } from '../../../../models/book.model';
 
 @Component({
   selector: 'app-book-detail',
