+++ angular/wolox-books/src/app/screens/auth/screens/book-detail/resolvers/get-book-detail.resolver.ts
@@ -7,13 +7,14 @@ import {
   Router
 } from '@angular/router';
 
-import { BookService } from './../../../services/book/book.service';
-import { Book } from './../../../models/book.model';
+import { BookService } from '../../../../../services/book/book.service';
+import { Book } from '../../../../../models/book.model';
 
 import 'rxjs/add/operator/catch';
+import 'rxjs/add/observable/of';
 
 @Injectable()
-export class GetBookListResolver implements Resolve<Observable<string | Book[]>> {
+export class GetBookDetailResolver implements Resolve<Observable<string | Book>> {
   constructor(
     private bookService: BookService,
     private router: Router
@@ -22,8 +23,9 @@ export class GetBookListResolver implements Resolve<Observable<string | Book[]>>
   resolve(
     route: ActivatedRouteSnapshot,
     state: RouterStateSnapshot
-  ): Observable<string | Book[]> {
-    return this.bookService.getBookList().catch(() => {
+  ): Observable<string | Book> {
+    return this.bookService.getBookDetail(route.params.id).catch(() => {
+      this.router.navigate(['book-list']);
       return Observable.of('data not available at this time');
     });
   }
