+++ angular/wolox-books/src/app/screens/auth/screens/books/books.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 
-import { Book } from './../../../../models/book.model';
+import { Book } from '../../../../models/book.model';
 
 @Component({
   selector: 'app-books',
