+++ angular/wolox-books/src/app/screens/auth/screens/books/resolvers/get-book-list.resolver.ts
@@ -7,14 +7,13 @@ import {
   Router
 } from '@angular/router';
 
-import { BookService } from './../../../services/book/book.service';
-import { Book } from './../../../models/book.model';
+import { BookService } from '../../../../../services/book/book.service';
+import { Book } from '../../../../../models/book.model';
 
 import 'rxjs/add/operator/catch';
-import 'rxjs/add/observable/of';
 
 @Injectable()
-export class GetBookDetailResolver implements Resolve<Observable<string | Book>> {
+export class GetBookListResolver implements Resolve<Observable<string | Book[]>> {
   constructor(
     private bookService: BookService,
     private router: Router
@@ -23,9 +22,8 @@ export class GetBookDetailResolver implements Resolve<Observable<string | Book>>
   resolve(
     route: ActivatedRouteSnapshot,
     state: RouterStateSnapshot
-  ): Observable<string | Book> {
-    return this.bookService.getBookDetail(route.params.id).catch(() => {
-      this.router.navigate(['book-list']);
+  ): Observable<string | Book[]> {
+    return this.bookService.getBookList().catch(() => {
       return Observable.of('data not available at this time');
     });
   }
